requirements.txt
aiohttp==3.9.1
aioitertools==0.11.0
aiosignal==1.3.1
amqp==2.6.1
aniso8601==9.0.1
apispec==6.3.0
astroid==2.15.5
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
billiard==3.6.4.0
blinker==1.6.2
boto3==1.26.142
botocore==1.29.142
cachecontrol==0.14.0
cachelib==0.9.0
cachetools==4.2.4
celery==4.4.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==41.0.1
decorator==5.1.1
dill==0.3.6
exceptiongroup==1.1.1
firebase-admin==6.4.0
Flask==2.3.2
flask-apispec==0.11.4
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
frozenlist==1.4.1
gevent==23.9.1
google-api-core==1.34.0
google-api-python-client==2.90.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.11.0
google-cloud-translate==3.12.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.59.3
grpcio-status==1.48.2
h11==0.14.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==4.6.11
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
marshmallow==3.19.0
mccabe==0.7.0
msgpack==1.0.7
multidict==6.0.4
packaging==23.1
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.0.0
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==3.20.3
psycopg2-binary==2.9.6
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pylint==2.17.4
pyparsing==3.1.0
pyTelegramBotAPI==4.12.0
pytest==7.3.1
python-dateutil==2.8.2
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3
PyYAML==6.0.1
redis==4.5.5
requests==2.31.0
retry==0.9.2
rsa==4.9
s3transfer==0.6.1
sentry-sdk==1.39.2
simple-websocket==1.0.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
SQLAlchemy-serializer==1.4.1
tomli==2.0.1
tomlkit==0.11.8
typing-extensions==4.6.2
uritemplate==4.1.1
urllib3==1.26.16
vine==1.3.0
wcwidth==0.2.6
webargs==8.2.0
Werkzeug==2.3.4
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.15.0
zope.event==5.0
zope.interface==6.1
